requirements.txt
pyscard>=2.0.0
keyboard>=0.13.5
pyusb>=1.2.1